+++ pkg/replicate/scheme_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) The Thanos Authors.
 // Licensed under the Apache License 2.0.
 
-package replicator
+package replicate
 
 import (
 	"bytes"
